requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
docutils==0.18.1
idna==3.3
Kivy==2.1.0
Kivy-Garden==0.1.5
Naked==0.1.31
plyer==2.0.0
pyasn1==0.4.8
pycryptodome==3.14.1
Pygments==2.11.2
pystache==0.6.0
PyYAML==6.0
requests==2.27.1
rsa==4.8
shellescape==3.8.1
static==1.1.1
urllib3==1.26.9
